requirements.txt
slither-analyzer>=0.10.4
torch_geometric>=2.5.0
torch>=2.2.0